# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export type TransactionInputMerchantCategory = typeof TransactionInputMerchantCategory[keyof typeof TransactionInputMerchantCategory];


export const TransactionInputMerchantCategory = {
  retail: 'retail',
  travel: 'travel',
  digital_goods: 'digital_goods',
  grocery: 'grocery',
  dining: 'dining',
  services: 'services',
} as const;

export type TransactionInputDeviceType = typeof TransactionInputDeviceType[keyof typeof TransactionInputDeviceType];


export const TransactionInputDeviceType = {
  mobile: 'mobile',
  desktop: 'desktop',
  tablet: 'tablet',
} as const;

export interface TransactionInput {
  /** @minimum 0.01 */
  transaction_amount: number;
  /**
     * @minimum 0
     * @maximum 23
     */
  transaction_hour: number;
  /**
     * @minimum 0
     * @maximum 6
     */
  day_of_week: number;
  merchant_category: TransactionInputMerchantCategory;
  /**
     * @minLength 2
     * @maxLength 2
     */
  country: string;
  device_type: TransactionInputDeviceType;
  is_new_device: boolean;
  /** @minimum 0 */
  transactions_last_1h: number;
  /** @minimum 0 */
  transactions_last_24h: number;
  /** @minimum 0 */
  avg_spend: number;
  /** @minimum 0 */
  spend_std_dev: number;
  /** @minimum 0 */
  account_age_days: number;
}

export type ReasonImpact = typeof ReasonImpact[keyof typeof ReasonImpact];


export const ReasonImpact = {
  positive: 'positive',
  negative: 'negative',
} as const;

export interface Reason {
  feature: string;
  impact: ReasonImpact;
  detail: string;
  contribution: number;
}

export type FraudPredictionPrediction = typeof FraudPredictionPrediction[keyof typeof FraudPredictionPrediction];


export const FraudPredictionPrediction = {
  HIGH_RISK: 'HIGH RISK',
  LOW_RISK: 'LOW RISK',
} as const;

export interface FraudPrediction {
  /**
     * @minimum 0
     * @maximum 1
     */
  fraud_probability: number;
  prediction: FraudPredictionPrediction;
  top_reasons: Reason[];
  scored_at: string;
  model_version: string;
}

export type TransactionRecordPrediction = typeof TransactionRecordPrediction[keyof typeof TransactionRecordPrediction];


export const TransactionRecordPrediction = {
  HIGH_RISK: 'HIGH RISK',
  LOW_RISK: 'LOW RISK',
} as const;

export type TransactionRecord = TransactionInput & {
  id: string;
  fraud_probability: number;
  prediction: TransactionRecordPrediction;
  created_at: string;
};

export interface CurvePoint {
  x: number;
  y: number;
}

export interface ModelMetrics {
  model_name: string;
  model_version: string;
  precision: number;
  recall: number;
  f1_score: number;
  roc_auc: number;
  pr_auc: number;
  confusion_matrix: number[][];
  roc_curve: CurvePoint[];
  pr_curve: CurvePoint[];
}

export type DashboardSummaryModelStatus = typeof DashboardSummaryModelStatus[keyof typeof DashboardSummaryModelStatus];


export const DashboardSummaryModelStatus = {
  READY: 'READY',
  TRAINING: 'TRAINING',
  DEGRADED: 'DEGRADED',
} as const;

export interface DashboardSummary {
  total_transactions: number;
  flagged_transactions: number;
  prevented_value: number;
  alert_rate: number;
  avg_probability: number;
  model_status: DashboardSummaryModelStatus;
  last_updated: string;
}

export type ListTransactionsParams = {
/**
 * @minimum 1
 * @maximum 20
 */
limit?: number;
};