# lib/api-zod/src/generated/types/dashboardSummary.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { DashboardSummaryModelStatus } from './dashboardSummaryModelStatus';

export interface DashboardSummary {
  total_transactions: number;
  flagged_transactions: number;
  prevented_value: number;
  alert_rate: number;
  avg_probability: number;
  model_status: DashboardSummaryModelStatus;
  last_updated: string;
}