# lib/api-zod/src/generated/types/fraudPrediction.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { FraudPredictionPrediction } from './fraudPredictionPrediction';
import type { Reason } from './reason';

export interface FraudPrediction {
  /**
     * @minimum 0
     * @maximum 1
     */
  fraud_probability: number;
  prediction: FraudPredictionPrediction;
  top_reasons: Reason[];
  scored_at: string;
  model_version: string;
}