# lib/api-zod/src/generated/types/modelMetrics.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { CurvePoint } from './curvePoint';

export interface ModelMetrics {
  model_name: string;
  model_version: string;
  precision: number;
  recall: number;
  f1_score: number;
  roc_auc: number;
  pr_auc: number;
  confusion_matrix: number[][];
  roc_curve: CurvePoint[];
  pr_curve: CurvePoint[];
}