# lib/api-zod/src/generated/types/reason.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ReasonImpact } from './reasonImpact';

export interface Reason {
  feature: string;
  impact: ReasonImpact;
  detail: string;
  contribution: number;
}