# lib/api-zod/src/generated/types/transactionInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { TransactionInputDeviceType } from './transactionInputDeviceType';
import type { TransactionInputMerchantCategory } from './transactionInputMerchantCategory';

export interface TransactionInput {
  /** @minimum 0.01 */
  transaction_amount: number;
  /**
     * @minimum 0
     * @maximum 23
     */
  transaction_hour: number;
  /**
     * @minimum 0
     * @maximum 6
     */
  day_of_week: number;
  merchant_category: TransactionInputMerchantCategory;
  /**
     * @minLength 2
     * @maxLength 2
     */
  country: string;
  device_type: TransactionInputDeviceType;
  is_new_device: boolean;
  /** @minimum 0 */
  transactions_last_1h: number;
  /** @minimum 0 */
  transactions_last_24h: number;
  /** @minimum 0 */
  avg_spend: number;
  /** @minimum 0 */
  spend_std_dev: number;
  /** @minimum 0 */
  account_age_days: number;
}