# lib/api-zod/src/generated/types/transactionRecord.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { TransactionInput } from './transactionInput';
import type { TransactionRecordPrediction } from './transactionRecordPrediction';

export type TransactionRecord = TransactionInput & {
  id: string;
  fraud_probability: number;
  prediction: TransactionRecordPrediction;
  created_at: string;
};